# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Papers Please System API
  description: API для системы управления УПК (пограничные пункты)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/users:
    get:
      tags:
      - user-controller
      operationId: getAllUsers
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseUserResponse"
    post:
      tags:
      - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/v1/users/{id}:
    get:
      tags:
      - user-controller
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - user-controller
      operationId: partialUpdateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/users/me:
    get:
      tags:
      - user-controller
      operationId: getCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    patch:
      tags:
      - user-controller
      operationId: partialUpdateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
components:
  schemas:
    UserRequest:
      required:
      - email
      - name
      - role
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 2
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - MIGRANT
          - INSPECTOR
          - BOSS
          - SECURITY
          - GOD
        upkId:
          type: string
    UserResponse:
      required:
      - email
      - id
      - name
      - role
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - MIGRANT
          - INSPECTOR
          - BOSS
          - SECURITY
          - GOD
        upkId:
          type: string
...
    UserRequestPartial:
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 2
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - MIGRANT
          - INSPECTOR
          - BOSS
          - SECURITY
          - GOD
        upkId:
          type: string
...
    PagedResponseUserResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
...
    Unit:
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT